results/requirements.txt
pandas
mysql-connector-python
mysqlclient ~= 2.1.0
SQLAlchemy ~= 1.4.31